+++ tests/data_asset/test_parameter_substitution.py
@@ -64,7 +64,7 @@ def test_parameter_substitution(single_expectation_custom_data_asset):
     assert result["result"]["details"]["expectation_argument"] == "upstream_dag_value"
     assert suite["evaluation_parameters"] == {
         "upstream_dag_key": "upstream_dag_value"}
-    assert suite["expectations"][0]["kwargs"] == {
+    assert suite.expectations[0].kwargs == {
         "expectation_argument": {"$PARAMETER": "upstream_dag_key"}}
 
 
@@ -79,7 +79,7 @@ def test_exploratory_parameter_substitution(single_expectation_custom_data_asset
     assert result["result"]["details"]["expectation_argument"] == "temporary_value"
     assert "evaluation_parameters" not in suite or suite["evaluation_parameters"] == {
     }
-    assert suite["expectations"][0]["kwargs"] == {
+    assert suite.expectations[0].kwargs == {
         "expectation_argument": {"$PARAMETER": "upstream_dag_key"}}
 
     # Evaluating the expectation without the parameter should now fail, because no parameters were set
@@ -92,7 +92,7 @@ def test_exploratory_parameter_substitution(single_expectation_custom_data_asset
     single_expectation_custom_data_asset.set_evaluation_parameter(
         "upstream_dag_key", "upstream_dag_value")
     validation_result = single_expectation_custom_data_asset.validate()
-    assert validation_result["results"][0]["result"]["details"]["expectation_argument"] == "upstream_dag_value"
+    assert validation_result.results[0]["result"]["details"]["expectation_argument"] == "upstream_dag_value"
 
 
 def test_validation_substitution(single_expectation_custom_data_asset):
@@ -105,7 +105,7 @@ def test_validation_substitution(single_expectation_custom_data_asset):
     # Provide a run-time evaluation parameter
     validation_result = single_expectation_custom_data_asset.validate(
         evaluation_parameters={"upstream_dag_key": "upstream_dag_value"})
-    assert validation_result["results"][0]["result"]["details"]["expectation_argument"] == "upstream_dag_value"
+    assert validation_result.results[0]["result"]["details"]["expectation_argument"] == "upstream_dag_value"
 
 
 def test_validation_substitution_with_json_coercion(single_expectation_custom_data_asset):
@@ -120,7 +120,7 @@ def test_validation_substitution_with_json_coercion(single_expectation_custom_da
     # Provide a run-time evaluation parameter
     validation_result = single_expectation_custom_data_asset.validate(
         evaluation_parameters={"upstream_dag_key": {"upstream_dag_value"}})
-    assert validation_result["results"][0]["result"]["details"]["expectation_argument"] == ["upstream_dag_value"]
+    assert validation_result.results[0]["result"]["details"]["expectation_argument"] == ["upstream_dag_value"]
 
     # Verify that the entire result object including evaluation_parameters is serializable
     assert validation_result["evaluation_parameters"]["upstream_dag_key"] == ["upstream_dag_value"]