+++ tests/datasource/test_sqlalchemy_datasource.py
@@ -66,7 +66,7 @@ def test_sqlalchemy_datasource_custom_data_asset(data_context, test_db_connectio
     )
     assert type(batch).__name__ == "CustomSqlAlchemyDataset"
     res = batch.expect_column_func_value_to_be("col_1", 1)
-    assert res["success"] is True
+    assert res.success is True
 
 
 def test_standalone_sqlalchemy_datasource(test_db_connection_string):
@@ -148,4 +148,4 @@ def test_sqlalchemy_source_templating(sqlitedb_engine):
                               generator.yield_batch_kwargs("test", col_name="animal_name")
                               )
     res = df.expect_column_to_exist("animal_name")
-    assert res["success"] is True
+    assert res.success is True