+++ docs/changelog/changelog.rst
@@ -1,5 +1,7 @@
 .. _changelog:
 
+* INTERNAL TYPES
+* Change data_asset get_data_asset_name and save_data_asset_name to be properties.
 
 0.8.3__develop
 -----------------